Builder-Layer-End/requirements.txt
# Image Refresh Agent - Python Dependencies

# Core async HTTP client
aiohttp==3.9.1
aiosignal==1.3.1
frozenlist==1.4.0
multidict==6.0.4
yarl==1.9.3

# YAML configuration
PyYAML==6.0.1

# Testing framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pylint==3.0.2

# Type hints
types-PyYAML==6.0.12.12
types-aiofiles==23.2.0.0

# Logging and monitoring
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.0
